Add p11tool and p11-kit support
This commit adds support to retrieve trusted certificates using
p11tool and p11-kit. This provides a better cross distro support,
however it requires the distro to support it, because of that
the path guest strategy is kept as a fallback. So the behavior
remains untouched for hosts not supporting p11-kit.

Another major benefit of p11-kit approach is that it enables to
read host certificates from a socket, this is specially relevant
in container contexts where one might want to use host certificates
inside a containerized application.

Signed-off-by: David Cassany <dcassany@suse.com>